fix(): Normalize options.maxSize and throw errors where appropriate
The documentation does not make it very clear the difference between
passing an integer and a string for maxSize, there also seems to be
quite a bit of confusion around whether or not to use `k` or `kb`
but there is no reason that `k`, `kb` and `kilobytes` can't all be
used interchangeably (as an example).

This pull request starts off by renaming the `getMaxSize` function
to something more appropriate `normalizeMaxSize`, it also updates
the `size` option that is provided to `getStream` to determine more
quickly if `maxSize` is undefined, and if so go ahead and just send
`null`.

Next it normalizes the `maxSize` option and determines if it can
return more quickly based upon the type of the data that is provided.
Such that if an integer is provided, do a small check and bit of math
and return it, if a string is provided, check to see if it was sent
in an already correctly formatted quantifier and return it, if
a longer quantifier was used, go ahead and parse it and assign the
correct quantifier for passing it to the file-stream-rotator package.

It also now throws errors in the appropriate places if the provided
option can not be parsed instead of just silently passing null to
the file-stream-rotator package and confusing users.